+++ src/job_queue/periodic_tasks.rs
@@ -30,6 +30,7 @@ use mitra_utils::datetime::days_before_now;
 
 use crate::mastodon_api::settings::helpers::{
     import_follows_task,
+    move_followers_task,
     ImporterJobData,
 };
 use crate::payments::{
@@ -215,6 +216,15 @@ pub async fn importer_queue_executor(
                     config,
                     db_client,
                     user_id,
+                    address_list,
+                ).await?;
+            },
+            ImporterJobData::Followers { user_id, from_actor_id, address_list } => {
+                move_followers_task(
+                    config,
+                    db_client,
+                    user_id,
+                    from_actor_id,
                     address_list,
                 ).await?;
             },